Rename files & platform detection
