+++ rotatescreen/__init__.py
@@ -1,4 +1,9 @@
-try:
-    from .displayWindows import *
-except ImportError:
-    from .displayLinux import *
+import platform
+
+system = platform.system()
+if system == "Windows":
+    from .display_windows import *
+elif system == "Linux":
+    from .display_linux import *
+else:
+    raise Exception(f"ERROR: Platform '{system}' is not yet supported!")
